refactor oracleProvider test (anvil to hardhat)
