+++ test/hardhat/arbitrum.spec.ts
@@ -1,12 +1,14 @@
-import { setChain } from '../utils'
-import { prepareMarketTest } from './testHelper'
+import { setChain } from './utils'
+import { prepareMarketTest } from './market/testHelper'
 import * as helpers from '@nomicfoundation/hardhat-network-helpers'
+import { pythSpec } from './oracle_provider/specs'
 describe('[arbitrum]', async () => {
   let initChainSnapshot: helpers.SnapshotRestorer
   let deps: any
+  const targetNetwork = 'arbitrum_goerli'
   before(async () => {
     console.log('set chain')
-    await setChain('arbitrum_goerli')
+    await setChain(targetNetwork)
     deps = await prepareMarketTest('arbitrum')
     initChainSnapshot = await helpers.takeSnapshot()
   })
@@ -14,7 +16,7 @@ describe('[arbitrum]', async () => {
   beforeEach(async () => {
     await initChainSnapshot.restore()
   })
-  const { feeSpec, lensSpec, liquidationSpec, liuqiditySpec, tradeSpec } = require('./specs')
+  const { feeSpec, lensSpec, liquidationSpec, liuqiditySpec, tradeSpec } = require('./market/specs')
 
   const getDeps = () => deps
   feeSpec(getDeps)
@@ -22,4 +24,5 @@ describe('[arbitrum]', async () => {
   liquidationSpec(getDeps)
   liuqiditySpec(getDeps)
   tradeSpec(getDeps)
+  pythSpec(targetNetwork)
 })
