+++ test/hardhat/mantle.spec.ts
@@ -1,12 +1,13 @@
-import { setChain } from '../utils'
-import { prepareMarketTest } from './testHelper'
+import { setChain } from './utils'
+import { prepareMarketTest } from './market/testHelper'
 import * as helpers from '@nomicfoundation/hardhat-network-helpers'
+import { agniTest } from './swap/agni.spec'
 describe('[mantle]', async () => {
   let initChainSnapshot: helpers.SnapshotRestorer
   let deps: any
+  const targetNetwork = 'mantle_testnet'
   before(async () => {
-    console.log('set chain')
-    await setChain('mantle_testnet')
+    await setChain(targetNetwork)
     deps = await prepareMarketTest('mantle')
     initChainSnapshot = await helpers.takeSnapshot()
   })
@@ -15,7 +16,8 @@ describe('[mantle]', async () => {
     await initChainSnapshot.restore()
   })
 
-  const { feeSpec, lensSpec, liquidationSpec, liuqiditySpec, tradeSpec } = require('./specs')
+  const { feeSpec, lensSpec, liquidationSpec, liuqiditySpec, tradeSpec } = require('./market/specs')
+  const { supraSpec, pythSpec } = require('./oracle_provider/specs')
 
   const getDeps = () => deps
   feeSpec(getDeps)
@@ -23,4 +25,7 @@ describe('[mantle]', async () => {
   liquidationSpec(getDeps)
   liuqiditySpec(getDeps)
   tradeSpec(getDeps)
+  supraSpec(targetNetwork)
+  pythSpec(targetNetwork)
+  agniTest()
 })
