+++ test/hardhat/oracle_provider/specs/index.ts
@@ -1,0 +1,2 @@
+export { spec as supraSpec } from './supra.spec'
+export { spec as pythSpec } from './pyth.spec'
