+++ test/hardhat/utils.ts
@@ -1,10 +1,9 @@
+import { reset } from '@nomicfoundation/hardhat-toolbox/network-helpers'
+import { Interface, Result } from 'ethers'
 import { deployments, ethers, getNamedAccounts } from 'hardhat'
 import { DeployOptions } from 'hardhat-deploy/types'
 import util from 'util'
 import { logDeployed } from './log-utils'
-import { Interface, Result, TopicFilter } from 'ethers'
-import { HardhatEthersSigner } from '@nomicfoundation/hardhat-ethers/signers'
-import { reset } from '@nomicfoundation/hardhat-toolbox/network-helpers'
 
 export async function deployContract<T>(
   contractName: string,
@@ -47,139 +46,6 @@ export async function hardhatErrorPrettyPrint<T>(method: () => Promise<T>): Prom
     }
     throw e.error || e
   }
-}
-
-export type BatchDeployParam = {
-  contract: string
-  signer: HardhatEthersSigner
-  args: any[][]
-}
-
-export async function batchDeploy(param: BatchDeployParam) {
-  const from = await param.signer.getAddress()
-  const factory = await ethers.getContractFactory(param.contract, param.signer)
-
-  const hre = await import('hardhat')
-  const rpcUrl = (hre.network.config as any).url
-  let id = 1
-
-  const txParams = await Promise.all(
-    param.args.map((args) => factory.getDeployTransaction(...args))
-  )
-
-  const gasRes = await fetch(rpcUrl, {
-    method: 'POST',
-    body: JSON.stringify(
-      txParams.map((txParam) => ({
-        method: 'eth_estimateGas',
-        params: [txParam],
-        id: id++,
-        jsonrpc: '2.0'
-      }))
-    ),
-    headers: { 'Content-Type': 'application/json' }
-  })
-  const gasDatas = await gasRes.json()
-  const gas = gasDatas.map((g: any) => g.result)
-
-  const body = []
-  for (let i = 0; i < txParams.length; i++) {
-    body.push({
-      method: 'eth_sendTransaction',
-      params: [{ ...txParams[i], from, gas: gas[i] }],
-      id: id++,
-      jsonrpc: '2.0'
-    })
-  }
-
-  const txHashs = await fetch(rpcUrl, {
-    method: 'POST',
-    body: JSON.stringify(body),
-    headers: { 'Content-Type': 'application/json' }
-  })
-
-  const txHashResults = await txHashs.json()
-
-  // wait mining
-  const lastTxHash = txHashResults[txHashResults.length - 1].result
-  while (true) {
-    await new Promise((resolve) => setTimeout(resolve, 1500))
-    if (await param.signer.provider.getTransactionReceipt(lastTxHash)) {
-      break
-    }
-  }
-
-  const deployed = await fetch(rpcUrl, {
-    method: 'POST',
-    body: JSON.stringify(
-      txHashResults.map((result: any) => ({
-        method: 'eth_getTransactionReceipt',
-        params: [result.result], // txHash
-        id: id++,
-        jsonrpc: '2.0'
-      }))
-    ),
-    headers: { 'Content-Type': 'application/json' }
-  })
-  const receipts = await deployed.json()
-  return receipts.map((r: any) => r.result.contractAddress)
-}
-
-export type BatchCallByFunctionNameParam = {
-  iface: Interface
-  from: string
-  to: string
-  functionName: string
-  data: any
-}
-
-export async function batchCallByFunctionName(
-  params: BatchCallByFunctionNameParam[]
-): Promise<Result[]> {
-  const batchCallParams = params.map((param) => ({
-    from: param.from,
-    to: param.to,
-    data: param.iface.encodeFunctionData(param.functionName, param.data)
-  }))
-
-  const results = await batchCall(batchCallParams)
-
-  const decodedData = []
-  for (let i = 0; i < params.length; i++) {
-    const param = params[i]
-
-    decodedData.push(param.iface.decodeFunctionResult(param.functionName, results[i]))
-  }
-  return decodedData
-}
-
-export type BatchCallParam = {
-  from: string
-  to: string
-  data: string
-}
-
-export async function batchCall(params: BatchCallParam[]) {
-  const hre = await import('hardhat')
-  const rpcUrl = (hre.network.config as any).url
-
-  let id = 1
-
-  const reqs = params.map((param) => ({
-    method: 'eth_call',
-    params: [param, 'latest'],
-    id: id++,
-    jsonrpc: '2.0'
-  }))
-
-  const res = await fetch(rpcUrl, {
-    method: 'POST',
-    body: JSON.stringify(reqs),
-    headers: { 'Content-Type': 'application/json' }
-  })
-  const datas = await res.json()
-
-  return datas.map((data: any) => data.result)
 }
 
 export type SupraPrice = {
